Switching to the V4 endpoints. (#31)
Switching to the V4 endpoints. The datafeed now has support for connection events. Wrappers for the other events are still needed.

There are a couple of custom edits to the Agent's ApiClient.cs file that are needed to support multi-part mime and allow filenames to be specified for attachments without using FileStreams.

Converting unit tests to a .NET core project