Switch to using NSwag for code generation (#36)
Swagger CodeGen uses FubarCoder's RestSharp.Portable library which is not thread safe. Since the typical paradigm is to share an HTTP client, it is easy to get into trouble when using multiple API classes.

NSwag generates code that uses Microsoft's HttpClient class directly which is thread safe its generated code will not have the same problems.

The generated API classes are now marked as internal to favor use of the wrapper classes that we've built.

ApiException added to wrap the SwaggerException thrown by the generated code.