Add content-type header to attachment segment in multipart request body (#43)
Symphony's Agent API requires the multipart segment for the file attachment to include a content-type header. This is not explicitly required in the YAML description of the API, the Swagger spec, or the HTTP RFC for file attachments in multipart sections. The RFC suggests having one, but does not require it. The code generator that we use, nswag, does not add this header so this patches the generated code to add a content-type header which makes the AgentApi happy.